better tests for compare/equality/identity/hash of playbook types
any type really, but used by playbook object tests atm
